Refactoring, expose realtime scheduler function
