+++ CMakeLists.txt
@@ -50,6 +50,7 @@ add_subdirectory(common)
 ## Library
 add_library(franka SHARED
   src/control_loop.cpp
+  src/control_tools.cpp
   src/control_types.cpp
   src/duration.cpp
   src/errors.cpp
