+++ include/franka/lowpass_filter.h
@@ -2,6 +2,7 @@
 // Use of this source code is governed by the Apache-2.0 license, see LICENSE
 #pragma once
 
+#include <array>
 #include <cmath>
 
 /**
