+++ src/control_loop.cpp
@@ -1,20 +1,15 @@
 // Copyright (c) 2017 Franka Emika GmbH
 // Use of this source code is governed by the Apache-2.0 license, see LICENSE
 #include "control_loop.h"
-#include "control_data_checker.h"
-#include "platform_type.h"
 
-#ifdef LIBFRANKA_WINDOWS
-#include <Windows.h>
-#else
-#include <pthread.h>
-#endif
-
+#include <algorithm>
 #include <cerrno>
 #include <cstring>
 #include <exception>
 #include <fstream>
 
+#include <franka/control_tools.h>
+#include <franka/control_types.h>
 #include <franka/exception.h>
 #include <franka/lowpass_filter.h>
 #include <franka/rate_limiting.h>
@@ -27,6 +22,35 @@
 using namespace std::string_literals;  // NOLINT(google-build-using-namespace)
 
 namespace franka {
+
+namespace {
+
+template <typename T, size_t N>
+inline void checkFinite(const std::array<T, N>& array) {
+  if (!std::all_of(array.begin(), array.end(), [](double d) { return std::isfinite(d); })) {
+    throw std::invalid_argument("Commanding value is infinite or NaN.");
+  }
+}
+
+inline void checkElbow(const std::array<double, 2>& elbow) {
+  checkFinite(elbow);
+  if (!isValidElbow(elbow)) {
+    throw std::invalid_argument(
+        "Invalid elbow configuration given! Only +1 or -1 are allowed for the sign of the 4th "
+        "joint.");
+  }
+}
+
+inline void checkMatrix(const std::array<double, 16>& transform) {
+  checkFinite(transform);
+  if (!isHomogeneousTransformation(transform)) {
+    throw std::invalid_argument(
+        "libfranka: Attempt to set invalid transformation in motion generator. Has to be column "
+        "major!");
+  }
+}
+
+}  // anonymous namespace
 
 template <typename T>
 constexpr research_interface::robot::Move::Deviation ControlLoop<T>::kDefaultDeviation;
@@ -43,10 +67,13 @@ ControlLoop<T>::ControlLoop(RobotControl& robot,
       limit_rate_(limit_rate),
       cutoff_frequency_(cutoff_frequency) {
   bool throw_on_error = robot_.realtimeConfig() == RealtimeConfig::kEnforce;
+  std::string error_message;
+  if (!setCurrentThreadToHighestSchedulerPriority(&error_message) && throw_on_error) {
+    throw RealtimeException(error_message);
+  }
   if (throw_on_error && !hasRealtimeKernel()) {
     throw RealtimeException("libfranka: Running kernel does not have realtime capabilities.");
   }
-  setCurrentThreadToRealtime(throw_on_error);
 }
 
 template <typename T>
@@ -271,55 +298,6 @@ void ControlLoop<CartesianVelocities>::convertMotion(
     command->valid_elbow = false;
     command->elbow_c = {};
   }
-}
-
-void setCurrentThreadToRealtime(bool throw_on_error) {
-#ifdef LIBFRANKA_WINDOWS
-  auto get_last_windows_error = []() -> std::string {
-    DWORD error_id = GetLastError();
-    LPSTR buffer = nullptr;
-    size_t size = FormatMessageA(
-        FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
-        nullptr, error_id, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPSTR)(&buffer), 0, nullptr);
-    return std::string(buffer, size);
-  };
-
-  if (!SetPriorityClass(GetCurrentProcess(), REALTIME_PRIORITY_CLASS)) {
-    throw RealtimeException("libfranka: unable to set priority for the process: "s +
-                            get_last_windows_error());
-  }
-
-  if (!SetThreadPriority(GetCurrentThread(), THREAD_PRIORITY_TIME_CRITICAL)) {
-    throw RealtimeException("libfranka: unable to set priority for the thread: "s +
-                            get_last_windows_error());
-  }
-#else
-  const int thread_priority = sched_get_priority_max(SCHED_FIFO);
-  if (thread_priority == -1) {
-    throw RealtimeException("libfranka: unable to get maximum possible thread priority: "s +
-                            std::strerror(errno));
-  }
-  sched_param thread_param{};
-  thread_param.sched_priority = thread_priority;
-
-  if (pthread_setschedparam(pthread_self(), SCHED_FIFO, &thread_param) != 0) {
-    if (throw_on_error) {
-      throw RealtimeException("libfranka: unable to set realtime scheduling: "s +
-                              std::strerror(errno));
-    }
-  }
-#endif
-}
-
-bool hasRealtimeKernel() {
-#ifdef LIBFRANKA_WINDOWS
-  return true;
-#else
-  std::ifstream realtime("/sys/kernel/realtime", std::ios_base::in);
-  bool is_realtime;
-  realtime >> is_realtime;
-  return is_realtime;
-#endif
 }
 
 template class ControlLoop<JointPositions>;
