+++ src/control_loop.h
@@ -14,9 +14,6 @@
 
 namespace franka {
 
-void setCurrentThreadToRealtime(bool throw_on_error);
-bool hasRealtimeKernel();
-
 template <typename T>
 class ControlLoop {
  public:
